Updated the add_nav_box() function to not need to include the nav-menu.php file.
The issue was that the 'nav_menu_meta_box_object' filter had a function added to it before
the 'admin_init' action, which is earlier than the wp-admin/includes/nav-menu.php file
is loaded. So applying that filter would cause WP to fall on its face, because it's trying to
run a function that hasn't been loaded up yet.
So just use my new custom filter, or just don't care about it: 'navception_nav_menu_meta_box_object'